move tests to scheduler folder and change tags
