+++ tests/test/util/test_scheduling.cpp
@@ -3,8 +3,6 @@
 #include "faabric_utils.h"
 #include "fixtures.h"
 
-#include <faabric/scheduler/Scheduler.h>
-#include <faabric/util/config.h>
 #include <faabric/util/func.h>
 #include <faabric/util/scheduling.h>
 
@@ -12,7 +10,7 @@ using namespace faabric::util;
 
 namespace tests {
 
-TEST_CASE("Test building scheduling decisions", "[util][scheduling-decisions]")
+TEST_CASE("Test building scheduling decisions", "[util]")
 {
     int appId = 123;
     int groupId = 345;
@@ -48,7 +46,7 @@ TEST_CASE("Test building scheduling decisions", "[util][scheduling-decisions]")
 }
 
 TEST_CASE("Test converting point-to-point mappings to scheduling decisions",
-          "[util][scheduling-decisions]")
+          "[util]")
 {
     int appId = 123;
     int groupId = 345;
@@ -94,284 +92,5 @@ TEST_CASE("Test converting point-to-point mappings to scheduling decisions",
     REQUIRE(actual.groupIdxs == expectedGroupIdxs);
     REQUIRE(actual.messageIds == expectedMessageIds);
     REQUIRE(actual.hosts == expectedHosts);
-}
-
-class SchedulingDecisionTestFixture : public SchedulerTestFixture
-{
-  public:
-    SchedulingDecisionTestFixture() { faabric::util::setMockMode(true); }
-
-    ~SchedulingDecisionTestFixture() { faabric::util::setMockMode(false); }
-
-  protected:
-    int appId = 123;
-    int groupId = 456;
-    std::string masterHost = faabric::util::getSystemConfig().endpointHost;
-
-    // Helper struct to configure one scheduling decision
-    struct SchedulingConfig
-    {
-        std::vector<std::string> hosts;
-        std::vector<int> slots;
-        int numReqs;
-        bool forceLocal;
-        SchedulingTopologyHint topologyHint;
-        std::vector<std::string> expectedHosts;
-    };
-
-    // Helper method to set the available hosts and slots per host prior to
-    // making a scheduling decision
-    void setHostResources(std::vector<std::string> registeredHosts,
-                          std::vector<int> slotsPerHost)
-    {
-        assert(registeredHosts.size() == slotsPerHost.size());
-        auto& sch = faabric::scheduler::getScheduler();
-        sch.clearRecordedMessages();
-
-        for (int i = 0; i < registeredHosts.size(); i++) {
-            faabric::HostResources resources;
-            resources.set_slots(slotsPerHost.at(i));
-            resources.set_usedslots(0);
-
-            sch.addHostToGlobalSet(registeredHosts.at(i));
-
-            // If setting resources for the master host, update the scheduler.
-            // Otherwise, queue the resource response
-            if (i == 0) {
-                sch.setThisHostResources(resources);
-            } else {
-                faabric::scheduler::queueResourceResponse(registeredHosts.at(i),
-                                                          resources);
-            }
-        }
-    }
-
-    // We test the scheduling decision twice: the first one will follow the
-    // unregistered hosts path, the second one the registerd hosts one.
-    void testActualSchedulingDecision(
-      std::shared_ptr<faabric::BatchExecuteRequest> req,
-      const SchedulingConfig& config)
-    {
-        auto& sch = faabric::scheduler::getScheduler();
-        SchedulingDecision actualDecision(appId, groupId);
-
-        // Set resources for all hosts
-        setHostResources(config.hosts, config.slots);
-
-        // The first time we request the scheduling decision, we will follow the
-        // unregistered hosts path
-        actualDecision = sch.publicMakeSchedulingDecision(
-          req, config.forceLocal, config.topologyHint);
-        REQUIRE(actualDecision.hosts == config.expectedHosts);
-
-        // Set resources again to reset the used slots
-        setHostResources(config.hosts, config.slots);
-
-        // The second time we request the scheduling decision, we will follow
-        // the registered hosts path
-        actualDecision = sch.publicMakeSchedulingDecision(
-          req, config.forceLocal, config.topologyHint);
-        REQUIRE(actualDecision.hosts == config.expectedHosts);
-    }
-};
-
-TEST_CASE_METHOD(SchedulingDecisionTestFixture,
-                 "Test basic scheduling decision",
-                 "[util][scheduling-decision]")
-{
-    SchedulingConfig config = {
-        .hosts = { masterHost, "hostA" },
-        .slots = { 1, 1 },
-        .numReqs = 2,
-        .forceLocal = false,
-        .topologyHint = SchedulingTopologyHint::NORMAL,
-        .expectedHosts = { masterHost, "hostA" },
-    };
-
-    auto req = batchExecFactory("foo", "bar", config.numReqs);
-
-    testActualSchedulingDecision(req, config);
-}
-
-TEST_CASE_METHOD(SchedulingDecisionTestFixture,
-                 "Test overloading all resources defaults to master",
-                 "[util][scheduling-decision]")
-{
-    SchedulingConfig config = {
-        .hosts = { masterHost, "hostA" },
-        .slots = { 1, 1 },
-        .numReqs = 3,
-        .forceLocal = false,
-        .topologyHint = SchedulingTopologyHint::NORMAL,
-        .expectedHosts = { masterHost, "hostA", masterHost },
-    };
-
-    auto req = batchExecFactory("foo", "bar", config.numReqs);
-
-    testActualSchedulingDecision(req, config);
-}
-
-TEST_CASE_METHOD(SchedulingDecisionTestFixture,
-                 "Test force local forces executing at master",
-                 "[util][scheduling-decision]")
-{
-    SchedulingConfig config = {
-        .hosts = { masterHost, "hostA" },
-        .slots = { 1, 1 },
-        .numReqs = 2,
-        .forceLocal = false,
-        .topologyHint = SchedulingTopologyHint::NORMAL,
-        .expectedHosts = { masterHost, "hostA" },
-    };
-
-    auto req = batchExecFactory("foo", "bar", config.numReqs);
-
-    SECTION("Force local off")
-    {
-        config.forceLocal = false;
-        config.expectedHosts = { masterHost, "hostA" };
-    }
-
-    SECTION("Force local on")
-    {
-        config.forceLocal = true;
-        config.expectedHosts = { masterHost, masterHost };
-    }
-
-    testActualSchedulingDecision(req, config);
-}
-
-TEST_CASE_METHOD(SchedulingDecisionTestFixture,
-                 "Test master running out of resources",
-                 "[util][scheduling-decision]")
-{
-    SchedulingConfig config = {
-        .hosts = { masterHost, "hostA" },
-        .slots = { 0, 2 },
-        .numReqs = 2,
-        .forceLocal = false,
-        .topologyHint = SchedulingTopologyHint::NORMAL,
-        .expectedHosts = { "hostA", "hostA" },
-    };
-
-    auto req = batchExecFactory("foo", "bar", config.numReqs);
-
-    testActualSchedulingDecision(req, config);
-}
-
-TEST_CASE_METHOD(SchedulingDecisionTestFixture,
-                 "Test scheduling decision with many requests",
-                 "[util][scheduling-decision]")
-{
-    SchedulingConfig config = {
-        .hosts = { masterHost, "hostA", "hostB", "hostC" },
-        .slots = { 0, 0, 0, 0 },
-        .numReqs = 8,
-        .forceLocal = false,
-        .topologyHint = SchedulingTopologyHint::NORMAL,
-        .expectedHosts = { masterHost, masterHost, masterHost, masterHost },
-    };
-
-    auto req = batchExecFactory("foo", "bar", config.numReqs);
-
-    SECTION("Even slot distribution across hosts")
-    {
-        config.slots = { 2, 2, 2, 2 };
-        config.expectedHosts = { masterHost, masterHost, "hostA", "hostA",
-                                 "hostB",    "hostB",    "hostC", "hostC" };
-    }
-
-    SECTION("Uneven slot ditribution across hosts")
-    {
-        config.slots = { 3, 2, 2, 1 };
-        config.expectedHosts = { masterHost, masterHost, masterHost, "hostA",
-                                 "hostA",    "hostB",    "hostB",    "hostC" };
-    }
-
-    SECTION("Very uneven slot distribution across hosts")
-    {
-        config.slots = { 1, 0, 0, 0 };
-        config.expectedHosts = {
-            masterHost, masterHost, masterHost, masterHost,
-            masterHost, masterHost, masterHost, masterHost
-        };
-    }
-
-    testActualSchedulingDecision(req, config);
-}
-
-TEST_CASE_METHOD(SchedulingDecisionTestFixture,
-                 "Test sticky pairs scheduling topology hint",
-                 "[util][scheduling-decision]")
-{
-    SchedulingConfig config = {
-        .hosts = { masterHost, "hostA" },
-        .slots = { 1, 1 },
-        .numReqs = 2,
-        .forceLocal = false,
-        .topologyHint = SchedulingTopologyHint::PAIRS,
-        .expectedHosts = { masterHost, "hostA" },
-    };
-
-    std::shared_ptr<faabric::BatchExecuteRequest> req;
-
-    SECTION("Test with hint we only schedule to new hosts pairs of requests")
-    {
-        config.expectedHosts = { masterHost, masterHost };
-        req = batchExecFactory("foo", "bar", config.numReqs);
-    }
-
-    SECTION("Test hint does not apply for master requests")
-    {
-        config.slots = { 0, 1 };
-        config.numReqs = 1;
-        config.expectedHosts = { "hostA" };
-        req = batchExecFactory("foo", "bar", config.numReqs);
-    }
-
-    SECTION("Test with hint we may overload remote hosts")
-    {
-        config.hosts = { masterHost, "hostA", "hostB" };
-        config.numReqs = 5;
-        config.slots = { 2, 2, 1 };
-        config.expectedHosts = {
-            masterHost, masterHost, "hostA", "hostA", "hostA"
-        };
-        req = batchExecFactory("foo", "bar", config.numReqs);
-    }
-
-    SECTION("Test with hint we still overload master if running out of slots")
-    {
-        config.hosts = { masterHost, "hostA" };
-        config.numReqs = 5;
-        config.slots = { 2, 2 };
-        config.expectedHosts = {
-            masterHost, masterHost, "hostA", "hostA", masterHost
-        };
-        req = batchExecFactory("foo", "bar", config.numReqs);
-    }
-
-    SECTION("Test hint with uneven slot distribution")
-    {
-        config.hosts = { masterHost, "hostA" };
-        config.numReqs = 5;
-        config.slots = { 2, 3 };
-        config.expectedHosts = {
-            masterHost, masterHost, "hostA", "hostA", "hostA"
-        };
-        req = batchExecFactory("foo", "bar", config.numReqs);
-    }
-
-    SECTION("Test hint with uneven slot distribution and overload")
-    {
-        config.hosts = { masterHost, "hostA", "hostB" };
-        config.numReqs = 6;
-        config.slots = { 2, 3, 1 };
-        config.expectedHosts = { masterHost, masterHost, "hostA",
-                                 "hostA",    "hostA",    "hostA" };
-        req = batchExecFactory("foo", "bar", config.numReqs);
-    }
-
-    testActualSchedulingDecision(req, config);
 }
 }
